Move components to own dir
